Rename adapter to 'criterion'
Update diagnostics namespace
Update adapter path
Update adatper name
